Huge success - Rayleigh sky model implemented.
Sky map displays the anticipated circular pattern around the sun.
The degree of polarization is calculated straightforwardly as Wikipedia suggests.
We have two methods that calculate the angle. One calculates the angle vector which is
useful for 3D visualization. This has been working for the past few commits.

The method that is working for the first time is calculating the absolute value of the angle.
This is done by finding the angle between the normal to the origin-observed-zenith plane and the
angle vector calculated with the previous method. A perculiarity here is that for points observed in the west hemisphere
where divided by the sun zenith line, we need to invert the angle. Not sure why.

Also adding a sky_sample script which repeats some of the code in sky_map (bad.) but aims to present a view of the sky
through the viewer's perspective.

All behave and inline tests pass.