Fixing a timezone related bug in the sky position calculation. Sky Map can now take a date argument (defaults to today)
The pysolar library has an archaic time zone management - get_altitude and get_azimuth
assume they are given the localized time, but not a UTC time with tzinfo attribute but
trully raw UTC time + DST value. So we do that and see that the days of DST change work better, and the sun position in the sky map looks correct in summer instead of -1 hour skewed.