Adding yaw rotation.
The sky model generator now has the concept of local coordinate system (internal) and
world coordinate system. All the coordinates it receives as input assumes are world coordinate,
and to convert them to local rotates with yaw. Then when returning converts back to world, rotating with -yaw.

The assumed facing direction is shown with red arrow in the sky map. To rotate by 10 degrees, press left or right arrow key.