Fix for sky model yaw calculations
The confusion came to be because it was not clear whether we are rotating
the camera or the world. Now we are set that red arrow points north and we
are rotating the camera, thus the observed points we pass it are in local camera coordinates.
The only thing that rotates with yaw is the local position of the sun.